float -> double changes

+++ Servers/Filters/vtkDesktopDeliveryServer.cxx
@@ -202,7 +202,7 @@ void vtkDesktopDeliveryServer::PreRenderProcessing()
       int i;
       for (rens->InitTraversal(), i = 0; (ren = rens->GetNextItem()); i++)
         {
-        float *viewport = ren->GetViewport();
+        double *viewport = ren->GetViewport();
         ren->SetViewport(viewport[0]*this->ImageReductionFactor,
           viewport[1]*this->ImageReductionFactor,
           viewport[2]*this->ImageReductionFactor,
